+++ src/universal/Object/ObjectTrait.php
@@ -60,6 +60,13 @@ trait ObjectTrait
         return $this->updatedAt;
     }
 
+    public function setUpdatedAt(?\DateTimeInterface $updatedAt): self
+    {
+        $this->updatedAt = $updatedAt;
+
+        return $this;
+    }
+
     public function getDeletedAt(): ?\DateTimeInterface
     {
         return $this->deletedAt;